+++ src/gameplay/characters/systems/mod.rs
@@ -1,3 +1,2 @@
 pub mod animation;
 pub mod movement;
-pub mod event;
